Working PID
After a few hours of messing around I wasn't able to get a PID controller working. The issue was that the we would set a target angle for the steering motor to try to reach. This worked if the steering correction factor was small.

However if a very large correction factor was calculated, this would cause the steering motor to stop accepting commands because it woudl try and reach an impossible target angle. This was fixed by using the Motor.track_target() functon.