Evaluating Sensors
We have spend a few hours testing ultra sonic and color sensors for
determining when nto turn. We have also switched the compass for a gyro
as it makes the dirving code simpler(we don't have to account for the
wrap around from 359 degrees to 0)

We logged the data over a few different test runs to test the reliablity
of the navigation using ultrasonics. We found that sometimes the ultra
sonic detects erroneous distance readings. Our working theory is that
this is caused by the geometery of the walls of the course(corners).

We have not worked out how to solve this issue. So we decided to
evaluate the color sensor in its place. We set up a simple function that
returns the squared distance between the current color vector reading
and a reference blue vector. If this lies within some threshold then we
know we are seeing blue and should turn left.

This however was not as reliable as we thought it would be. Sometimes
the squared distance is not close to our threshold. This is probably due
to different lighting conditions causing false negatives with in our
threshold check. We will need to come up with a better method.

Another issue we encountered was the compass sensor. Its very appealing
because it gives absolute heading. However this is annoying as it would
be more convenient to have relative heading. Implemententing this would
be time consuming so we decided to evaluate the gyro sensor. We found
that it is fairly reliable but the drift due to integration can become
significant.

This is also why having 2 ultrasonic sensors is convenient. As driving
in a straight line becomes trivial and would never drift.