Testing Gryo
Pi Pico speaks to mpu6050 gyro and prints the z axis gyro value and the delta time. Python script reads this in over serial and integrates to get the current angle. Drift is very very minimal. However it can still even be improved by doing the integration on the pi pico and the only sending the current angle when a command comes in over serial.