More FFT analysis
We upgraded the code that plots the frequency domain for the gyro. From this we were able to determing a good cutoff frequency for gyro noise(roughly 600 Hz). We are using a dynamic lowpass filter on the pico with the RC constant we determined here.

The 2 plots in the logs folder showcases different cutoff frequencies and the gyro response