Project restructure
Restructured the repository to be more inline with what we need  for the
FE comp.

Also seperated the code in src into 2 directories. One for ev3 and one
for the pico. This helps with organisation as well as with working on
the ev3 code.

Prior to this, the vscode extension we were using was uploading the
entire project workspace(around 177 files) each time we ran main.py.

This made iterating painfully slow. Now work will be faster thanks to
restructuring and mvoing log files out of the repository. The data plots
might getadded back in to the documentation later on.