Finalized open challenge code and added commments
The open challenge code is now completed and working.
Robot can drive 3 laps in both directions and will stop
in its starting section. There are issues with the UART
interface when communication between pico and ev3 so we might
regress to using USB

Comments was added to every file explaining classes, functions,
variables as well as certain calculations. This is the beginning of
documentation