Updated color detection code
Color detection is now done asynchronously on the pico. If the pico detects blue or orange, it stores that until the ev3 tells it to reset the detected color. This prevents the issue of missing the colored lines due to the ev3 logic loop being too slow to poll the color sensor